refactor(cli): extract doctor command module
